pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-fastapi-app"
version = "0.1.0"
description = "Config-driven RAG API with FastAPI, Dynaconf, CLI, logging, and tests"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "CRAG Demo" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "dynaconf>=3.2.5",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.1",
  "typer>=0.12.3",
  "rich>=13.7.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "ruff>=0.6.3",
  "mypy>=1.11.2"
]

[project.scripts]
rag-cli = "app.cli:app"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --disable-warnings"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
